feat(rawdata): raw data from api (minor updates)
